moving lots of earth phase 1
