+++ src/components/ApplicationBlock/_ApplicationBlock.scss
