--- src/css/modules/_GlobalSearch.scss
@@ -1,26 +1,0 @@
-.coa-GlobalSearch {
-  input {
-    border: 1px solid $color-white;
-  }
-  button {
-    font-size: 20px !important;
-    background-color: $color-white;
-    color: $color-black;
-    border-left: 2px solid $color-gray-light;
-  }
-}
-
-.coa-GlobalSearch__button {
-  background-image: none !important;
-  &:hover {
-    svg path {
-      fill: $color-white;
-    }
-  }
-}
-
-.coa-GlobalSearch__icon {
-  @include media($coa-small-screen) {
-    display: none;
-  }
-}
