--- src/css/modules/_Hero.scss
@@ -1,22 +1,0 @@
-.coa-Hero {
-  @include box-verticalcenter(17rem);
-  background-color: $color-gray-lighter;
-  background-position: center center;
-  background-size: cover;
-  margin-bottom: $coa-space-level4;
-  padding: $coa-space-level5;
-
-  @include media($coa-small-screen) {
-    height: 24rem;
-  }
-  @include media($coa-medium-screen) {
-    height: 32rem;
-  }
-}
-
-.coa-Hero__callout {
-  h2 {
-    font-weight: $font-bold;
-    margin: 0;
-  }
-}
