+++ src/js/modules/Tile.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import I18nLink from 'js/modules/I18nLink'
+import I18nLink from 'components/I18nLinks/I18nLink'
 import ChevronCircleRightSVG from 'js/svg/ChevronCircleRight';
 
 const Tile = ({ url, text, tag }) => (
