+++ src/js/page_sections/Header.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { defineMessages, injectIntl } from 'react-intl';
 
 import Menu from 'js/page_sections/Menu/Menu';
-import I18nLink from 'js/modules/I18nLink';
-import ExternalLink from 'js/modules/ExternalLink';
+import I18nLink from 'components/I18nLinks/I18nLink';
+import ExternalLink from 'components/ExternalLink/ExternalLink';
 
 const i18nMessages = defineMessages({
   headerMenuButton: {
