+++ src/stories/Contact.stories.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import ContactMap from 'js/modules/Contact/ContactMap';
-import ContactDetails from 'js/modules/Contact/ContactDetails';
+import ContactMap from 'components/Contact/ContactMap';
+import ContactDetails from 'components/Contact/ContactDetails';
 import ContactPhone from 'js/modules/Contact/Phone';
 import ContactHours from 'js/modules/Contact/Hours';
 import ContactEmail from 'js/modules/Contact/Email';
