+++ src/stories/ExternalLink.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import ExternalLink from 'js/modules/ExternalLink';
+import ExternalLink from 'components/ExternalLink/ExternalLink';
 
 storiesOf('ExternalLink', module)
   .add('ExternalLink', () => (
